+++ iminds-api/src/main/java/com/stl/iminds/candidate/resource/CandidateSkillsDTO.java
@@ -14,7 +14,7 @@ public class CandidateSkillsDTO implements Serializable {
 	
 	private String description;
 	
-	private Long experience;
+	private Double experience;
 
 	/**
 	 * @return the candidateskillid
@@ -75,14 +75,14 @@ public class CandidateSkillsDTO implements Serializable {
 	/**
 	 * @return the experience
 	 */
-	public Long getExperience() {
+	public Double getExperience() {
 		return experience;
 	}
 
 	/**
 	 * @param experience the experience to set
 	 */
-	public void setExperience(Long experience) {
+	public void setExperience(Double experience) {
 		this.experience = experience;
 	}
